Split Gemma3nAudio.swift into focused modules and guard processor stub
  - Gemma3nAudioAttention.swift: relative position embeddings, chunked attention
  - Gemma3nAudioNorm.swift: cumulative group normalization
  - Gemma3nAudioConv.swift: SSCP conv blocks, subsampling projection
  - Gemma3nAudio.swift: conformer blocks, top-level encoder
  - Gemma3nVLM.swift: preconditionFailure on unimplemented processor stub
